Update Feature.php, add a setProperty to be able to modify a property
In the process or parsing data and creating a Feature I often reuse properties arrays to shorten amount of code.
But if I have 8 or 10 properties that are the same for many features in my map, I usually need a "name" property to be unique.
This is the best way I found to do this. Updating a property value in a single line of code when I need it.